+++ tests/test_core.py
@@ -14,7 +14,7 @@ import ndonnx.additional as nda
 from ndonnx import _data_types as dtypes
 from ndonnx._utility import promote
 
-from .utils import get_numpy_array_api_namespace, run
+from .utils import assert_array_equal, get_numpy_array_api_namespace, run
 
 
 def numpy_to_graph_input(arr, eager=False):
@@ -65,7 +65,7 @@ def test_make_graph_input(dtype, input, shape):
     a = ndx.array(shape=shape, dtype=dtype)
     model = ndx.build({"a": a}, {"b": a})
     actual = run(model, {"a": input})
-    np.testing.assert_equal(input, actual["b"])
+    assert_array_equal(actual["b"], input)
 
 
 def test_null_promotion():
@@ -77,19 +77,57 @@ def test_null_promotion():
         "b": np.array([4.0, 5.0, 6.0, 7.0]),
     }
     actual = run(model, inputs)
-    np.testing.assert_equal(np.add(inputs["a"], inputs["b"]), actual["c"])
+    assert_array_equal(actual["c"], np.add(inputs["a"], inputs["b"]))
 
 
-def test_asarray():
-    a = ndx.asarray([1, 2, 3], dtype=ndx.int64)
-    assert a.dtype == ndx.int64
-    np.testing.assert_array_equal(
-        np.array([1, 2, 3], np.int64), a.to_numpy(), strict=True
-    )
+@pytest.mark.parametrize(
+    "array, dtype, expected_dtype",
+    [
+        ([1, 2, 3], ndx.int64, ndx.int64),
+        (np.array([1, 2, 3], np.int64), None, ndx.int64),
+        (1, ndx.int64, ndx.int64),
+        (1, ndx.float64, ndx.float64),
+        (["a", "b"], ndx.utf8, ndx.utf8),
+        (np.array(["a", "b"]), None, ndx.utf8),
+        (np.array(["a", "b"], object), None, ndx.utf8),
+        ("a", ndx.utf8, ndx.utf8),
+        (np.array("a"), None, ndx.utf8),
+        (np.array("a", object), None, ndx.utf8),
+        ([["a"]], None, ndx.utf8),
+        (np.array([["a"]]), None, ndx.utf8),
+        (np.array([["a"]], object), None, ndx.utf8),
+    ],
+)
+def test_asarray(array, dtype, expected_dtype):
+    a = ndx.asarray(array, dtype=dtype)
+    assert a.dtype == expected_dtype
+    assert_array_equal(np.array(array, expected_dtype.to_numpy_dtype()), a.to_numpy())
 
 
-def test_asarray_masked():
-    np_arr = np.ma.masked_array([1, 2, 3], mask=[0, 0, 1])
+@pytest.mark.parametrize(
+    "np_arr",
+    [
+        np.ma.masked_array([1, 2, 3], mask=[0, 0, 1]),
+        np.ma.masked_array(1, mask=0),
+        np.ma.masked_array(["a", "b"], mask=[1, 0]),
+        np.ma.masked_array("a", mask=0),
+    ],
+)
+def test_asarray_masked(np_arr):
+    ndx_arr = ndx.asarray(np_arr)
+    assert isinstance(ndx_arr, ndx.Array)
+    assert isinstance(ndx_arr.to_numpy(), np.ma.MaskedArray)
+    assert_array_equal(np_arr, ndx_arr.to_numpy())
+
+
+@pytest.mark.parametrize(
+    "np_arr",
+    [
+        np.ma.masked_array(["a", "b"], mask=[1, 0], dtype=object),
+        np.ma.masked_array("a", mask=0, dtype=object),
+    ],
+)
+def test_asarray_object_dtype(np_arr):
     ndx_arr = ndx.asarray(np_arr)
     assert isinstance(ndx_arr, ndx.Array)
     assert isinstance(ndx_arr.to_numpy(), np.ma.MaskedArray)
@@ -101,14 +139,14 @@ def test_basic_eager_add(_a, _b):
     b = ndx.asarray(_b)
     x = ndx.asarray([1]) + 2
     c = a + b + x
-    np.testing.assert_array_equal(c.to_numpy(), _a + _b + np.array([1]) + 2)
+    assert_array_equal(c.to_numpy(), _a + _b + np.array([1]) + 2)
 
 
 def test_string_concatenation():
     a = ndx.asarray(["a", "b", "c"])
     b = ndx.asarray(["d", "e", "f"])
-    np.testing.assert_array_equal((a + b).to_numpy(), np.array(["ad", "be", "cf"]))
-    np.testing.assert_array_equal((a + "d").to_numpy(), np.array(["ad", "bd", "cd"]))
+    assert_array_equal((a + b).to_numpy(), np.array(["ad", "be", "cf"]))
+    assert_array_equal((a + "d").to_numpy(), np.array(["ad", "bd", "cd"]))
 
 
 def test_combining_lazy_eager():
@@ -134,7 +172,7 @@ def test_lazy_array(_a, _b):
     model = ndx.build({"a": a}, {"c": c})
     actual = run(model, {"a": _a})["c"]
     expected = _a * _b
-    np.testing.assert_equal(actual, expected)
+    assert_array_equal(actual, expected)
 
 
 def test_indexing(_a):
@@ -144,7 +182,7 @@ def test_indexing(_a):
     model = ndx.build({"a": a}, {"b": b})
     actual = run(model, {"a": _a})["b"]
     expected = _a[0]
-    np.testing.assert_array_equal(expected, actual)
+    assert_array_equal(expected, actual)
 
 
 @pytest.mark.parametrize(
@@ -160,6 +198,11 @@ def test_indexing(_a):
             np.ma.masked_array([1, 2, 3], mask=[0, 0, 1]),
             np.array([3.12, 3.24, -124.0]),
         ),
+        (
+            np.array([True, False, True], dtype=np.bool_),
+            np.ma.masked_array([1, 2, 3], mask=[0, 0, 1]),
+            np.ma.masked_array([3.12, 3.24, -124.0], mask=[0, 0, 0]),
+        ),
     ],
 )
 def test_where(condition, x, y):
@@ -172,7 +215,8 @@ def test_where(condition, x, y):
         {"condition": condition_onnx, "x": x_onnx, "y": y_onnx}, {"out": out}
     )
     actual = run(model, {"condition": condition, "x": x, "y": y})["out"]
-    np.testing.assert_array_equal(expected, actual)
+    # NumPy simply drops the masked array in `np.where`. We do not want to do the same.
+    np.testing.assert_array_equal(actual, expected)
 
 
 def test_indexing_on_scalar():
@@ -185,7 +229,7 @@ def test_indexing_on_scalar():
 def test_indexing_on_scalar_mask():
     res = ndx.asarray([])
     res = res[False]
-    np.testing.assert_equal(nda.shape(res).to_numpy(), (0, 0))
+    assert_array_equal(nda.shape(res).to_numpy(), (0, 0))
 
 
 def test_indexing_with_mask(_a):
@@ -199,7 +243,7 @@ def test_indexing_with_mask(_a):
 
     model = ndx.build({"a": a, "mask_": mask}, {"c": c})
     actual = run(model, {"a": _a, "mask_": _mask})["c"]
-    np.testing.assert_array_equal(expected_c, actual)
+    assert_array_equal(expected_c, actual)
 
 
 def test_indexing_with_mask_raises(_a):
@@ -221,7 +265,7 @@ def test_indexing_with_array(_a):
 
     model = ndx.build({"a": a, "index_": index}, {"c": c})
     actual = run(model, {"a": _a, "index_": _index})["c"]
-    np.testing.assert_array_equal(expected_c, actual)
+    assert_array_equal(expected_c, actual)
 
 
 def test_indexing_with_tuple_of_array(_a):
@@ -241,7 +285,7 @@ def test_slicing(_a):
 
     model = ndx.build({"a": a}, {"b": b})
     actual = run(model, {"a": _a})["b"]
-    np.testing.assert_equal(actual, _a[1:2])
+    assert_array_equal(actual, _a[1:2])
 
 
 def test_indexing_list(_a):
@@ -258,7 +302,7 @@ def test_indexing_slice_ellipsis(_c):
     model = ndx.build({"a": a}, {"b": b})
     actual = run(model, {"a": _c})["b"]
     expected = _c[..., 1]
-    np.testing.assert_array_equal(actual, expected)
+    assert_array_equal(actual, expected)
 
 
 def test_indexing_none(_c):
@@ -270,7 +314,7 @@ def test_indexing_none(_c):
     model = ndx.build({"a": a}, {"b": b})
     actual = run(model, {"a": _c})["b"]
     expected = _c[_index]
-    np.testing.assert_array_equal(actual, expected)
+    assert_array_equal(actual, expected)
 
 
 def test_illegal_indexing(_a):
@@ -309,7 +353,7 @@ def test_indexing_set_with_array(_a):
 
     model = ndx.build({"a": a, "index_": index}, {"c": c})
     actual = run(model, {"a": _a, "index_": _index})["c"]
-    np.testing.assert_equal(actual, expected_c)
+    assert_array_equal(actual, expected_c)
 
 
 def test_indexing_set_scalar():
@@ -323,7 +367,7 @@ def test_indexing_set_scalar():
     expected_c = 0
 
     model = ndx.build({"a": a}, {"c": c})
-    np.testing.assert_array_equal(expected_c, run(model, {"a": _a})["c"], strict=True)
+    assert_array_equal(expected_c, run(model, {"a": _a})["c"])
 
 
 # Check if repr does not raise
@@ -349,7 +393,7 @@ def test_rops(rop):
     b = ndx.asarray([1.0, 2.0, 3.0])
 
     res = getattr(a, rop)(b)
-    np.testing.assert_equal(res.to_numpy(), getattr(a.to_numpy(), rop)(b.to_numpy()))
+    assert_array_equal(res.to_numpy(), getattr(a.to_numpy(), rop)(b.to_numpy()))
 
 
 def test_rsub():
@@ -358,7 +402,7 @@ def test_rsub():
     res = 1 - a
     a_value = a.to_numpy()
     assert a_value is not None
-    np.testing.assert_equal(res.to_numpy(), 1 - a_value)
+    assert_array_equal(res.to_numpy(), 1 - a_value)
 
 
 def test_matrix_transpose():
@@ -367,9 +411,9 @@ def test_matrix_transpose():
 
     model = ndx.build({"a": a}, {"b": b})
     npx = get_numpy_array_api_namespace()
-    np.testing.assert_equal(
-        npx.matrix_transpose(npx.reshape(npx.arange(3 * 2 * 3), (3, 2, 3))),
+    assert_array_equal(
         run(model, {"a": np.arange(3 * 2 * 3, dtype=np.int64).reshape(3, 2, 3)})["b"],
+        npx.matrix_transpose(npx.reshape(npx.arange(3 * 2 * 3), (3, 2, 3))),
     )
 
 
@@ -381,9 +425,9 @@ def test_matrix_transpose_attribute():
     npx = get_numpy_array_api_namespace()
     expected = npx.reshape(npx.arange(3 * 2 * 3), (3, 2, 3)).mT
 
-    np.testing.assert_equal(
-        expected,
+    assert_array_equal(
         run(model, {"a": np.arange(3 * 2 * 3, dtype=np.int64).reshape(3, 2, 3)})["b"],
+        expected,
     )
 
 
@@ -392,7 +436,7 @@ def test_transpose_attribute():
     b = a.T
 
     model = ndx.build({"a": a}, {"b": b})
-    np.testing.assert_equal(
+    assert_array_equal(
         np.reshape(np.arange(3 * 2), (3, 2)).T,
         run(model, {"a": np.arange(3 * 2, dtype=np.int64).reshape(3, 2)})["b"],
     )
@@ -408,51 +452,51 @@ def test_array_spox_interoperability():
         np.arange(3 * 2, dtype=np.int64).reshape(3, 2), mask=np.ones((3, 2), dtype=bool)
     )
     actual = run(model, {"a": input})["b"]
-    np.testing.assert_equal(expected, actual)
+    assert_array_equal(actual, expected)
 
 
 def test_creation_arange():
     a = ndx.arange(0, stop=10)
-    np.testing.assert_equal(a.to_numpy(), np.arange(stop=10))
+    assert_array_equal(a.to_numpy(), np.arange(stop=10))
 
     b = ndx.arange(1, 10)
-    np.testing.assert_equal(b.to_numpy(), np.arange(1, 10))
+    assert_array_equal(b.to_numpy(), np.arange(1, 10))
 
     c = ndx.arange(1, 10, 2)
-    np.testing.assert_equal(c.to_numpy(), np.arange(1, 10, 2))
+    assert_array_equal(c.to_numpy(), np.arange(1, 10, 2))
 
     d = ndx.arange(0.0, None, step=-1)
-    np.testing.assert_array_equal(
-        np.arange(0.0, None, step=-1), d.to_numpy(), strict=True
-    )
+    assert_array_equal(np.arange(0.0, None, step=-1), d.to_numpy())
 
 
 def test_creation_full():
     a = ndx.full((2, 3), 5)
-    np.testing.assert_equal(np.full((2, 3), 5), a.to_numpy())
+    assert_array_equal(a.to_numpy(), np.full((2, 3), 5))
 
     b = ndx.full((2, 3), 5, dtype=ndx.float32)
-    np.testing.assert_equal(np.full((2, 3), 5, dtype=np.float32), b.to_numpy())
+    assert_array_equal(b.to_numpy(), np.full((2, 3), 5, dtype=np.float32))
     c = ndx.full((2, 3), "a", dtype=ndx.nutf8)
-    np.testing.assert_equal(np.full((2, 3), "a"), c.to_numpy())
+    assert_array_equal(
+        c.to_numpy(), np.ma.masked_array(np.full((2, 3), "a"), mask=False)
+    )
 
     d = ndx.full(2, 5, dtype=ndx.int8)
-    np.testing.assert_equal(np.full(2, 5, dtype=np.int8), d.to_numpy())
+    assert_array_equal(d.to_numpy(), np.full(2, 5, dtype=np.int8))
 
     # Check lazy creation
     e = ndx.array(shape=tuple(), dtype=ndx.int64)
     f = ndx.full(e, 10)
     model_proto = ndx.build({"e": e}, {"f": f})
     actual = run(model_proto, {"e": np.array(5, dtype=np.int64)})["f"]
-    np.testing.assert_equal(np.array([10] * 5, dtype=np.int64), actual)
+    assert_array_equal(actual, np.array([10] * 5, dtype=np.int64))
 
     # Note we must know the output shape to export an ONNX artifact.
     g = ndx.array(shape=(2,), dtype=ndx.int64)
     h = ndx.full(g, 10)
     model_proto = ndx.build({"g": g}, {"h": h})
-    np.testing.assert_equal(
-        np.array([[10, 10, 10], [10, 10, 10]]),
+    assert_array_equal(
         run(model_proto, {"g": np.array([2, 3], dtype=np.int64)})["h"],
+        np.array([[10, 10, 10], [10, 10, 10]]),
     )
 
 
@@ -475,14 +519,14 @@ def test_result_type(args, expected):
 
 def test_ceil():
     a = ndx.asarray(np.array([1.2, 1.3, -13.13]))
-    np.testing.assert_equal(ndx.ceil(a).to_numpy(), [2.0, 2.0, -13.0])
+    assert_array_equal(ndx.ceil(a).to_numpy(), [2.0, 2.0, -13.0])
 
 
 def test_propagates_minimal_dtype():
     a = ndx.asarray([1, 2, 4], dtype=ndx.int8)
     b = a + 1
     assert b.dtype == ndx.int8
-    np.testing.assert_equal(b.to_numpy(), [2, 3, 5])
+    assert_array_equal(b.to_numpy(), np.array([2, 3, 5], dtype=np.int8))
 
 
 @pytest.mark.parametrize(
@@ -496,7 +540,7 @@ def test_propagates_minimal_dtype():
     ],
 )
 def test_all(x):
-    np.testing.assert_equal(ndx.all(x).to_numpy(), np.all(x.to_numpy()))
+    assert_array_equal(ndx.all(x).to_numpy(), np.all(x.to_numpy()))
 
 
 @pytest.mark.parametrize(
@@ -514,7 +558,7 @@ def test_searchsorted(side):
     a = ndx.asarray(a_val, dtype=ndx.int64)
     b = ndx.asarray(b_val, dtype=ndx.int64)
     c = ndx.searchsorted(a, b, side=side)
-    np.testing.assert_equal(c_val, c.to_numpy())
+    assert_array_equal(c_val, c.to_numpy())
 
 
 @pytest.mark.skip(reason="TODO: onnxruntime")
@@ -536,7 +580,7 @@ def test_searchsorted_nans(side):
 
     model = ndx.build({"a": a, "b": b}, {"c": c})
 
-    np.testing.assert_equal(c_val, run(model, dict(a=a_val, b=b_val))["c"])
+    assert_array_equal(c_val, run(model, dict(a=a_val, b=b_val))["c"])
 
 
 def test_searchsorted_raises():
@@ -558,7 +602,7 @@ def test_truediv():
     y = ndx.asarray([2, 3, 3], dtype=ndx.int64)
     z = x / y
     assert isinstance(z.dtype, ndx.Floating)
-    np.testing.assert_array_equal(z.to_numpy(), np.array([0.5, 2 / 3, 1.0]))
+    assert_array_equal(z.to_numpy(), np.array([0.5, 2 / 3, 1.0]))
 
 
 @pytest.mark.parametrize(
@@ -574,6 +618,24 @@ def test_truediv():
         ndx.nint16,
         ndx.nint32,
         ndx.nint64,
+    ],
+)
+def test_prod(dtype):
+    x = ndx.asarray([2, 2]).astype(dtype)
+    y = ndx.prod(x)
+    if isinstance(dtype, ndx.Nullable):
+        input = np.asarray([2, 2], dtype=dtype.values.to_numpy_dtype())
+        input = np.ma.masked_array(input, mask=False)
+    else:
+        input = np.asarray([2, 2], dtype=dtype.to_numpy_dtype())
+    actual = np.prod(input)
+
+    assert_array_equal(y.to_numpy(), actual)
+
+
+@pytest.mark.parametrize(
+    "dtype",
+    [
         ndx.uint8,
         ndx.uint16,
         ndx.uint32,
@@ -582,11 +644,19 @@ def test_truediv():
         ndx.nuint32,
     ],
 )
-def test_prod(dtype):
+def test_prod_unsigned(dtype):
+    # We intentionally deviate from the Array API standard and reduce for <= uint32
+    # using uint32 as our default unsigned type due to lack of kernel support for uint64
     x = ndx.asarray([2, 2]).astype(dtype)
     y = ndx.prod(x)
+    if isinstance(dtype, ndx.Nullable):
+        input = np.asarray([2, 2], dtype=dtype.values.to_numpy_dtype())
+        input = np.ma.masked_array(input, mask=False)
+    else:
+        input = np.asarray([2, 2], dtype=dtype.to_numpy_dtype())
+    actual = np.prod(input).astype(np.uint32)
 
-    np.testing.assert_array_equal(y.to_numpy(), np.array(4))
+    assert_array_equal(y.to_numpy(), actual)
 
 
 @pytest.mark.parametrize(
@@ -652,14 +722,10 @@ def test_numerical_unary_operations_fail_on_non_numeric_input(operation, dtype):
 def test_string_shape_operations():
     a = ndx.asarray(["a", "b", "c"])
     b = ndx.broadcast_to(a, (2, 3))
-    np.testing.assert_array_equal(
-        b.to_numpy(), np.array([["a", "b", "c"], ["a", "b", "c"]])
-    )
+    assert_array_equal(b.to_numpy(), np.array([["a", "b", "c"], ["a", "b", "c"]]))
 
     c = ndx.concat([a, a], axis=0)
-    np.testing.assert_array_equal(
-        c.to_numpy(), np.array(["a", "b", "c", "a", "b", "c"])
-    )
+    assert_array_equal(c.to_numpy(), np.array(["a", "b", "c", "a", "b", "c"]))
 
 
 @pytest.mark.parametrize(
@@ -671,9 +737,7 @@ def test_string_shape_operations():
 )
 def test_zeros(dtype):
     x = ndx.zeros((2, 3), dtype=dtype)
-    np.testing.assert_equal(
-        x.to_numpy(), np.zeros((2, 3), dtype=dtype.to_numpy_dtype())
-    )
+    assert_array_equal(x.to_numpy(), np.zeros((2, 3), dtype=dtype.to_numpy_dtype()))
 
 
 @pytest.mark.xfail(reason="https://github.com/onnx/onnx/issues/6276")
@@ -681,7 +745,7 @@ def test_empty_concat_eager():
     a = ndx.asarray([], ndx.int64)
     b = ndx.asarray([1, 2, 3], ndx.int64)
     out = ndx.concat([a, b], axis=0)
-    np.testing.assert_equal(out.to_numpy(), b.to_numpy())
+    assert_array_equal(out.to_numpy(), b.to_numpy())
 
 
 def test_empty_concat_lazy_known_shape():
@@ -695,7 +759,7 @@ def test_empty_concat_lazy_known_shape():
 
     out = run(model, {"a": anp, "b": bnp})["out"]
 
-    np.testing.assert_equal(out, bnp)
+    assert_array_equal(out, bnp)
 
 
 def test_empty_concat_lazy_unknown_shape():
@@ -709,7 +773,7 @@ def test_empty_concat_lazy_unknown_shape():
 
     out = run(model, {"a": anp, "b": bnp})["out"]
 
-    np.testing.assert_equal(out, bnp)
+    assert_array_equal(out, bnp)
 
 
 # if the precision loss looks concerning, note https://data-apis.org/array-api/latest/API_specification/type_promotion.html#mixing-arrays-with-python-scalars
@@ -755,6 +819,10 @@ def test_promotion_failures(arrays, scalar):
         promote(*arrays, scalar)
 
 
+@pytest.mark.skipif(
+    np.__version__ <= "1",
+    reason="Cross kind scalar promotion not specified in NumPy < 2",
+)
 @pytest.mark.parametrize(
     "x, y",
     [
@@ -774,4 +842,87 @@ def test_cross_kind_promotions(x, y):
     if isinstance(y, np.ndarray):
         y = ndx.asarray(y)
     onnx_result = x + y
-    np.testing.assert_equal(onnx_result.to_numpy(), np_result)
+    assert_array_equal(onnx_result.to_numpy(), np_result)
+
+
+@pytest.mark.parametrize(
+    "x, y, expected",
+    [
+        ([0], [], []),
+        ([[0], [0]], [0, 0, 0], [[0, 0, 0], [0, 0, 0]]),
+        ([1, 2, 3],) * 3,
+        ([],) * 3,
+    ],
+)
+@pytest.mark.parametrize(
+    "x_dtype, y_dtype, expected_dtype",
+    [
+        (ndx.int64,) * 3,
+        (ndx.nint64,) * 3,
+        (ndx.float64,) * 3,
+        (ndx.nfloat64,) * 3,
+        (ndx.int64, ndx.nint64, ndx.nint64),
+        (ndx.float64, ndx.nfloat64, ndx.nfloat64),
+        (ndx.nint64, ndx.float64, ndx.nfloat64),
+    ],
+)
+def test_where_equal_arrays(x, y, expected, x_dtype, y_dtype, expected_dtype):
+    cond = ndx.array(shape=(), dtype=ndx.bool)
+    x = ndx.asarray(x).astype(x_dtype)
+    y = ndx.asarray(y).astype(y_dtype)
+
+    result = ndx.where(cond, x, y)
+
+    assert result.dtype == expected_dtype
+    # NumPy simply drops the masked array in `np.where`. We do not want to do the same.
+    np.testing.assert_array_equal(result.to_numpy(), expected)
+
+
+@pytest.mark.parametrize(
+    "x, expected_shape",
+    [
+        (ndx.array(shape=(1, 2), dtype=ndx.utf8), (1, 2)),
+        (ndx.array(shape=(1, 2), dtype=ndx.nuint8), (1, 2)),
+        (ndx.array(shape=(1, None, 5), dtype=ndx.nuint8), (1, None, 5)),
+        (ndx.array(shape=("N", "M"), dtype=ndx.float64), (None, None)),
+        (
+            ndx.reshape(
+                ndx.array(shape=(4,), dtype=ndx.utf8),
+                ndx.array(shape=(1,), dtype=ndx.int64),
+            ),
+            (None,),
+        ),
+        (
+            ndx.reshape(ndx.array(shape=(4,), dtype=ndx.utf8), ndx.asarray([2, 2])),
+            (2, 2),
+        ),
+    ],
+)
+def test_lazy_array_shape(x, expected_shape):
+    assert x.shape == expected_shape
+
+
+@pytest.mark.parametrize(
+    "x, shape",
+    [
+        (
+            ndx.array(shape=("N",), dtype=ndx.utf8),
+            ndx.array(shape=("M",), dtype=ndx.int64),
+        ),
+        (
+            ndx.array(shape=("N", 1), dtype=ndx.utf8),
+            ndx.array(shape=("N"), dtype=ndx.int64),
+        ),
+        (
+            ndx.array(shape=(1,), dtype=ndx.utf8),
+            ndx.array(shape=("N"), dtype=ndx.int64),
+        ),
+        (
+            ndx.array(shape=(), dtype=ndx.utf8),
+            ndx.array(shape=("N"), dtype=ndx.int64),
+        ),
+    ],
+)
+def test_dynamic_reshape_has_no_static_shape(x, shape):
+    with pytest.raises(ValueError, match="Could not determine static shape"):
+        ndx.reshape(x, shape).shape